Run tests as part of CI process
We want to fail the build if any of the tests fail.

It still makes sense to validate whether we can build the site, even if the tests fail, so that the developer can get as much feedback as possible. By defining the `rspec` command as a separate script, Travis will continue and run the `middleman build` regardless of the exit status, but will fail the build if either script fails.