Use the same Sass lint config as GOV.UK Frontend
For now we need to copy this config file as there is no other way to share it. We should revisit this in the future to prevent the two from drifting apart.

Unlike the frontend repo we are not defining paths in config yet as it is a simpler app and we're not sure that it's needed.