Add shim to support HTML5 elements in IE 6-9
This is the existing HTML 5 shim[1] as used in govuk_template[2], which enables use of HTML5 sectioning elements in legacy Internet Explorer and provides basic HTML5 styling for Internet Explorer 6-9, Safari 4.x (and iPhone 3.x), and Firefox 3.x.

This also adds support for Sprockets asset packaging, which allows us to follow thr same conventions as govuk_template when including vendored scripts.

We should revisit how this shim is included when we have a better understanding of what we're going to do with the various scripts (including this one) that are currently bundled in govuk_template.

[1]: https://github.com/aFarkas/html5shiv
[2]: https://github.com/alphagov/govuk_template/blob/e9ae7ff045e0adf552aeed509be5ef174be325aa/source/assets/javascripts/vendor/html5shiv-printshiv.js