Add basic auth for private beta
The authentication exists to ensure that users do not start using the Design System instead of the current resources - there is still a lot of work to do to get to a place where we’d want this to be happening.

There’s no need to keep the credentials ‘secret’ as they’re only there as a deterrent, which is why it’s OK to commit the encrypted htpasswd details to a public git repository.