Update markup to use Frontend 0.19
Remove `govuk-u-font-smoothing` classes as this is now handled by `govuk-typography-common`.

Remove override classes where a specific component/partial class is in use and move override styles into that component/partial.

Change references to `breadcrumb` components to `breadcrumbs`. Also remove reference to breadcrumb from mobile-navigation as this is no longer used in the partial.