Override `prose-scope` link styles
As per FE 0.0.22-alpha, the `govuk-link` styles are now applied to `a` tags
within prose scope.

These styles have higher specificity than the app styles due to the way the
prose scope uses HTMl selectors. This interferes with some of the styles of the
copy link and the tabs.

This overrides those interfering styles by adding more specific classes to the
component styles.