Reset font weight for nav & mobile nav links
The font weight set in the new `govuk-link` style resets the font weight of
links within the nav components.

The bold font weight is set in `app-c-navigation`, and the `li` of the mobile
nav which no longer cascades to the link.

We should possibly look into refactoring the SASS so that the two components
at least set their font styles consistently.