Auto-load GOV.UK Frontend macros
Glob the file system for macro file within the @govuk-frontend node modules folder, parse each file and grab any macros they export, and then make them available as globals.

This allows us to remove the `{% from “component/macro.njk" import govukComponent %}` lines from the examples we create, which means that Prototype Kit users will not have to worry about what to do with that line.

Because we need a Nunjucks environment we can use to parse the templates, the config for the nunjucks engine has been moved into the config folder and required in both places.

We should be able to adopt this in the Prototype Kit as well (though we’ll probably need to swap the spread operators out for something that works in older versions of Node as well).