Add frontendComponentName parameter to example macro
In teh case of text-input there is no such component in Frontend.
There is an Input component. To accommodate such instances,
an optional ` frontendComponentName` parameter has been added to the macro.
If it is present then lin kto table of arguments will use that. If not,
then `params.item` will be used as it's set as default.