Fix colour swatch styles for IE8
By default static breakpoints is set to desktop
and those media queries get rasterized.

For colour swatches we have a custom breakpoint that
is heigher than desktop so IE8 never gets those styles.

To fix, we check if $govuk-is-ie8 == true and if true
we set the custom breakpoint to match desktop.