Handle form submissions by redirecting to referrer
Handle form submissions from examples that contain forms by redirecting them back to the referring example. This relies on updating the form examples within the design system to POST to `/form-handler`.

If for whatever reason there is no referrer, then redirect to the homepage instead.

Note that this will only work when deployed to CloudFoundry – Netlify does not use the nginx config at all.