Do not track users who have enabled 'DoNotTrack'
Adds a guard around Google Tag Manager script to prevent it from initialising,
for users who have indicated they do not want to be tracked.

Also removes the `<noscript></noscript>` include as we can't guard this.

If we were server rendering the site, we could take advantage of the 'DoNotTrack' header.