Refer to non-gov users as ‘public‘ not ‘citizens‘
I think that citizens is a problematic word because:

- the opposite of a citizen is someone who is stateless, which is not
  the distinction we’re trying to draw here
- people who aren’t citizens sometimes need to use government services,
  for example those applying for asylum

I think ‘public’ is a better word – we are public servants,
delivering public services.

***

The reason for making this distinction at all is because we design
differently for people who use a service repeatedly. I don’t think that 
having a catch-all word for these users is realistic because there’s a
wide variety of such users. In individual services it’s probably better
to refer to them by the role or task they’re employed to do, for example
case worker, solicitor, team leader. But ‘non-public’ works to
differentiate these users when talking generally.