Remove unneccessary feature detects from Modernizr
We don’t use the flexboxtweener or history detection within the Design System – I believe this is a hangover from when we initially used the tech-docs template as a base, which does use both features.

Flexboxtweener is used to support Flexbox in IE10, but we have opted to fall back to a non-flexbox layout in that browser instead.

The history API detection is used in the tech docs to update the hash when scrolling.