Remove HTML5 print shiv from Modernizr
We bundle up an HTML5 shiv in our IE8 javascript file by using the html5shiv module directly, so there’s no need to include one here as well.

Technically the HTML5 shiv is used in IE9, but IE9 supports semantic elements natively, with the exception of the `<main>` element which needs styling in IE10 and IE11 as well - so we’re not relying on this.