Add ‘example wrapper’ for full page examples
When generating the site, there are two nunjucks render calls for each page. The _content_ of each individual page is converted from nunjucks to HTML, and *then* that HTML is inserted into the nunjucks layout in a separate call. This means that we can’t target different blocks from a layout (e.g. beforeContent, content) from the examples.

To get around this, we move the parts of the template we want to be able to target (beforeContent, content and the mainClasses) into a new nunjucks template which 'full page' examples can then extend. When the content of the page is generated we end up with a `<div class="govuk-width-container">` which we can insert directly into the `main` block of GOV.UK Frontend.

This has the added benefit that the HTML we generate for the code tabs will include the layout wrappers, but not the entire boilerplate, header and footer. The downside is that we do end up with some duplication of parts of the template from Frontend.