Move caption out of headings in typography section
In most cases, the caption is an orientation device and should not form part of the H1.

For example, in our ‘question pages’ the captions we use are ‘Question 3 of 9’ and ‘About you’ – including these in the H1 would make the page headings ‘Question 3 of 9 Your details’ and ‘About you What is your home adddress?’ neither if which really make sense.

There will be some cases where including the caption in the H1 _does_ make sense, but we think these will be an exception.