Use captions appropriately in Design System
It makes sense to include the caption within the page heading for the themes ‘Ask users for…’ and ‘Help users to…’ but not for ‘Pages’ nor when we’re presenting the category (‘Styles’, ‘Components’, ‘Getting started’)

For this reason, this updates the template to only include the caption as part of the page heading if the theme contains an ellipsis. It also removes the ellipsis as part of the caption.

As an example:

- the page ‘Addresses’ in the ‘Ask users for…’ theme would have the H1 ‘Asking users for Addresses’, where ‘Asking users for’ is styled as a caption.
- the page ‘Confirmation pages’ in the ‘Pages’ theme would have the H1 ‘Confirmation pages’ with a separate caption ‘Pages’.
- the page ‘Buttons’ in the ‘Components’ section would have the H1 ‘Buttons’ with a separate caption ‘Components’