Create document store when indexing
The search index only contains what's needed to match and identify a document, but won't give us back anything other than the document's identifier (`ref`).

We need to create a ‘document store’ which will allow us to lookup the information about the document that we can use to present the result.

Co-authored-by: Nick Colley <nick.colley@digital.cabinet-office.gov.uk>