Configure Metalsmith to rollup search.js
And to concatenate it into application.js
We've had to go down this route instead of the approach
that's in out other files because
`metalsmith-assets` was copying over the .html file from
node_modules/lunr and the `metalsmith-broken-link-checker` was
was failing on that file.

By using rollup we've also avoided having to copy over
accessible autocomplete/dist javascript file.