Add styles for search components
Add styles for the container, text input and search results
Import site-search.scss partial into main.scss file

Ideally we'd use an inline-svg image for the search icon.

Unfortunately on slow connections the search icon would appear before
accessible auto complete initialised. This is becuase we can only
control the svg visability with `js-enabled`
and `.no-js` classes which are added before the javascript executes.

We are also unabled to place the svg inside the
`app-site-search__wrapper` which is generated by the Accessible autocomeplete (AA).
We can however set the SVG as a background image of that container as
DATA-URI, preventing the image to appear before the search box.
As AA doesn't support IE8, the backgroud doesn't appear and DATA-URIs
are supported in IE9 and up.