Allow closing the open section in mobile nav
You cannot currently close the section of the page you are currently in in the mobile nav.

The sections are hidden by a `--hidden` modifier which is applied to every section except the current section. This modifier is never removed – it’s just overruled by the `--active` modifier which the javascript toggles. However, because the current section does not have the `--hidden` modifier, removing the `--active` modifier has no effect.

To fix this, we remove the `--hidden` modifier entirely and make all sections hidden by default, with the visibility then controlled by the `--active` modifier only. We also have to add the `--active` modifier to the toggle, so that it is displayed with the extra top border as designed.