Fingerprint all CSS files
Update the config to fingerprint all CSS files, and use the fingerprint helper to include them.

Because we specify some asset paths using relative URLs (e.g. `../grid-nested-annotate.css`), but the fingerprint object has everything relative to the root, we need to ‘normalise’ all URLs using the path. Unfortunately that means we need to know the path of the page including the asset, and we don’t know that within metalsmith-in-place (because metalsmith-in-place runs before metalsmith-permalinks). This means we can call getFingerprint with relative URLs in layouts, but not in the source files themselves.

For that reason, the ‘block areas’ example has been updated to specify a path relative to the root.