Use simpler default example for text inputs
Change the default example for the text input component from an input asking for an email address to an example asking for someone’s full name.

As raised in the backlog [1], we are not currently using the correct `type=“email”`. We could fix this, but as users are likely to take this ‘default’ example and modify it to ask different questions, it makes sense to make it as generic as possible, avoiding using a type specific to a particular type of input.

Therefore changing to a more ‘generic’ question, such as full name, which does not require specific input types, makes more sense.

[1]: https://github.com/alphagov/govuk-design-system-backlog/issues/51#issuecomment-414986345