Disable browser form validation
Most modern browsers provide some sort of validation for fields that have explicit types (such as email, or telephone) or when a field is marked as required. The way that this validation works is inconsistent across browsers, and we have little control over the content or the style of the error messages that are used. This means that there are multiple ‘stages’ of validation going on, and there is no consistency between them.

For this reason, we have previously recommended disabling the browser’s native validation (though this doesn’t seem to be written down anywhere, yet) so we should do this in our own examples.

For full page examples that already have forms, it’s as simple as adding the novalidate attribute to that form. For other examples, they don’t currently have a form tag, so we have to add one to the example layout. This does mean that all examples, including examples that don’t use any form elements, will be wrapped in a form, but I can’t currently think of any problems this would cause. Because the form tag is part of the layout, it won’t appear in the code blocks in the Design System.