Fix monospace font size
Major web browsers automatically reduce the size of monospace text
in a range of situations.

They do this due to the greater width of many monospace typefaces
in comparison to other fonts at the same text height.

While well-intentioned, their handling of monospace fonts is
fundamentally broken, and an unusual CSS rule is required to fix
the problem.