Make search test more robust
When adding different content that feeds search it flagged that tests
were not set up correctly to match how the search matching logic works.

Updated tests in now test for words beginning with the query in the
title (words) and then separately in the aliases, correctly accounting
for the fact that there may not be aliases present.