Add test to verify adding page titles to search index
This tests the extraction of page titles into metadata and using that
data to generate search index entries.

To test this a new fixtures has been added that contains markdown syntax

Extract-page-headings and metalsmith-in-place plugins have been added to
allow for headings to be extracted (as extraction can only be done
from markdown files) and page to be compiled into .html so that the lunr
plugin can generate the index.