Use location.href instead of location.pathname
Page heading entries contain a hash (#) in the string for the path
property to allow navigation to the exact location on the page.

Using location.pathname with a path containing a hash, causes the hash
to be encoded as  %23.

While we could create a separate urlFragment property for those
store entries, check for it when navigate to the search result url,
this change is simpler and has no downsides.