Add data attributes to track contact links
We already track the equivalent links on the homepage, this just applies the same data attributes to support links in the panel at the bottom of most pages and on the ‘Get in touch’ page.

This will help us to understand what channels users go to for support.

Also removes duplicated class attributes from links on the get in touch page.