Render options as table markup
Split the primary options and other component options/ nested options into separate tables following
https://www.w3.org/WAI/tutorials/tables/two-headers/#table-with-header-cells-in-the-top-row-and-first-column

The alternative https://www.w3.org/WAI/tutorials/tables/multi-level/
of keeping all options in one table would have required each cell to reference
three other cells using `headers`. W3C recommends the first option over this:
https://www.w3.org/WAI/tutorials/tables/multi-level/#split-up-multi-level-tables

Add an “options-“ span element that we link to display the table if it’s in the hash. We can’t link to details element itself as it’s hidden on page load inside the tabs.

This doesn't use GOV.UK Frontend table macro as it doesn't accept HTML generated in the template (adding a `caller` to the table macro would fix this - we have a card in the backlog).

Move GTM tag “github-component-arguments” to details element (we’ll change the tag to be more descriptive later as part of the analytics tidy up).