Only build once in CI
Travis already does a a build of the Design System to test that it builds correctly, so if we do that first then when running in the context of CI we can create a basic server serving the contents of deploy/public without having to do a second build. This looks for the existing of a CI environment variable set to true, which should be present on Travis. [1]

This should make our tests quicker to run on CI (because we only perform one ‘build’), and prevent jest failing because of a server start timeout (as it should always take less than 30 seconds to spin up a server if we don’t also need to do a build!)

[1]: https://docs.travis-ci.com/user/environment-variables/#default-environment-variables