Update package lock file
In a7adde9ad48c68a4e1aff82aca48abdc7e84a25d we removed gulp and gulp-sass-lint from our list of dependencies, but the corresponding changes to the lock file were not included (that commit includes a change to the lock file, but they’re irrelevant changes to the `resolved` properties of some of the packages.

This commit attempts to bring things back in line, by running `npm remove gulp gulp-sass-lint` and committing the resulting lock file.