Split out brand colour; use brand-colour variable
Rather than referencing the blue from the palette directly, use the brand colour and split it out from the ‘status’ section.

Removes a reference to the phase banner, as the colour page should only list colours that are used globally, rather than the colours for every component. For example, we don’t currently list turquoise despite it being used by the panel component.

By doing this we make the distinction between applications of colour and the colour palette clearer.