Remove fixed layout
In a previous accessibility audit we had reports of issues with the sticky and independently scrollable parts of our original design.

We also need to meet the new guidance in WCAG 2.1 which means that the design system reflows well when zoomed: https://www.w3.org/TR/WCAG21/#reflow

This change turns the design system into a static layout with a maximum width, which means there is more vertical space for reading, and reflows better without causing obstructions.

Additionally this change makes the design system less 'special' and more like the services that our users build. It also allows us to present the examples in the design system (For example, a component) at a width which is closer to how they will actually be used in a service.

We have created a new back to top component that will return you to the navigation, since some of our pages are long and no longer have the navigation available as you scroll.