Fix grammar error in 'When to use this component'
CHANGE:

You can use it to display metadata like ‘Last updated’ with a date like ‘22 June 2018’. Or to summarise a user’s responses at the end of a form like the [check answers](/patterns/check-answers) pattern.

TO:

You can use it to display metadata like ‘Last updated’ with a date like ‘22 June 2018’, or to summarise a user’s responses at the end of a form like the [check answers](/patterns/check-answers) pattern.

REASON:

Incorrect grammar.