Fix bug with invalid example IDs
This reverts a hotfix https://github.com/alphagov/govuk-design-system/pull/809
that removed the changes in favour of a proper fix for
https://github.com/alphagov/govuk-design-system/issues/620

Generating example IDs from example titles can be brittle if there
are unexpected characters in titles. Use slugger for more robust converting of
exampleTitle to exampleId. This also uses the solution from #620 to fix
duplicate IDs by adding an extra parameter "titleSuffix" to examples if
necessary.

Co-authored-by: Nick Colley <nick.colley@digital.cabinet-office.gov.uk>