Use h2 instead of h3 for column heading
Discussed this with Alistair and we agreed that h2 would be better so we
don't have a jump from h1 to h3. We also identified that the real issue
was that the opening tag had been updated but not the closing tag which
was the cause of the confusion between using h2 or h3.