Improve links to Slack channels
Updates the existing Slack links to use Slack's 'deep link' redirect. These will try to open the channel in the user's Slack app if it's installed, whilst providing links to download the app and to open in browser to help users if it's not.

This allows us to remove the 'open in app' links which were flagged as part of a recent audit as it does not make sense out of context.

> “As I navigated out of context, I discovered a link which read to me as “Open in app.” It was not immediately clear to me while situated out of context which app the link related to as this was not obvious within the link text. Although I could establish the purpose of the link in context in a testing environment some users might not be aware of which app is intended to open. Including the app name within the link text will enable users to clearly understand what will occur if the link is selected. This was present using JAWS, NVDA, VoiceOver for iPhone and Mac”

It also makes the way we describe channels consistent, and adds a full stop to the end of the sentence in the get in touch panel, because grammar.