Avoid unnecessary 301 redirects
The links in the navigation currently omit the trailing slash. For example, the 'Components' link links to /components.

If you try to visit this URL you get a 301 to /components/.

By adding the trailing slashes to the links in the navigation, we save a round trip to the server to get the 301 response.