Add smoke tests for CloudFoundry deploys
This creates a new HTML page which will be built and deployed to `/__canary__/index.html` containing nothing but the words ‘Tweet tweet’ (following the pattern set by GOV.UK [1])

It adds a smoke-test script which verifies the existence and content of that file, returning an error exit code if is not present.

Finally, it tells the Blue/Green deploy script that we use to use that smoke test, which means if e.g. nginx fails to start the deploy will not continue, and the existing container and route will be maintained.

This should hopefully give us more confidence when rolling out changes to the CloudFoundry / nginx configuration.

[1]: https://www.gov.uk/__canary__