Add explicit routes to the manifest file
When the blue green deploy plugin runs, it’s picking design-system.service.gov.uk as the first route and using that to generate the fully qualified domain name used for smoke tests.

This means that smoke tests are running against govuk-design-system-origin-new.design-system.service.gov.uk which is not valid.

Specifying the routes in the manifest (and, importantly, putting the cloudapps.digital domain first) should fix this.

This behaviour was not seen in testing as we were not testing against an instance with a cdn-route.

Co-authored-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>