Update label for mobile navigation
The ID referenced by `aria-labelledby` is incorrect – there is no element with this ID on the page.

I think it’s trying to reference the ‘Menu’ button in the header, but this has the ID `app-mobile-nav-toggler` (without ‘header’).

We could update the reference to use this ID, but given that the ‘Menu’ button ends up with an aria-label of ‘Toggle mobile menu’, the navigation would end up with the label ‘Toggle mobile menu navigation’. This is not the correct label for the navigation landmark itself – it only makes sense on the toggle button.

Instead, we provide a textual aria-label of ‘Top-level’. As per best practise, we omit the term "navigation", as the screen reader will read both the role and the contents of the label (e.g. Voiceover reads ‘Top-level navigation’)