Update label on mobile menu button
This updates the label on the menu button in the header so that it uses the same term as the navigation landmark (‘top level navigation’ rather than ‘mobile menu’).

We can also make it clearer what the button does by changing the label depending on the state of the navigation – rather than always labelling the button with ‘Toggle top-level navigation’ we can use ‘Show top-level navigation’ or ‘Hide top-level navigation’ depending on whether the menu is currently collapsed.