Fix missing aria on mobile category navigation
Since the focus state changes in 031500a, there is now an extra `<div>` wrapping the `a.js-mobile-nav-subnav-toggler`. This means we now need to ascend two levels up the DOM tree before we can look for the associated navigation.

This change was made in the click handler (which means we are updating aria-expanded when the sub-navigation opens and closes) but the initial setup fails, which means parts of the sub-navigation that have not yet been interacted with do not have any ARIA attributes, and the aria-label is never applied.