Improve label for sub-navigation sections
Now that the ARIA attributes are being applied correctly, the lists of links within each section are being labelled-by their toggle.

The issue with this is that the toggle itself has an aria-label of e.g. ‘Show pages within Components’ which means that the section ends up with this label too. This causes e.g. Voiceover to announce ‘’Show pages within Components, group’ when you focus the first link within each section.

We can address this by removing the aria-labelledby association with the toggle, and adding an explicit aria-label with the section name. Because this section can never be reached without JS enabled (and because arguably the ARIA markup makes sense regardless of the expand/collapse functionality) we can add this in the markup, rather than enhancing it with JavaScript.