Simplify ARIA markup
Remove redundant labels on each category ‘sub nav’ heading, as the expanded/collapsed state is already conveyed through the use of the `aria-expanded` attribute.

Move the `aria-controls` and `aria-label` on to the toggle button markup, as this element should be hidden in a non-javascript context, so if the toggle is visible the association and label should always make sense.