Use inputmode="numeric" and pattern for relevant fields
We've done some investigation to show that now in certain
scenarios this is the best way of collecting numeric data from users. We have a
bunch of changes to make to roll this out in GOV.UK Frontend:
https://github.com/alphagov/govuk-frontend/issues/1449#issuecomment-504006087
After that, we can set this in the input component itself instead of using the
`attributes` object.

Following our above investigation, GOV.UK Pay introduced this change for
collecting credit card details:
https://github.com/alphagov/pay-frontend/pull/921 They haven't seen
any issues so we are sufficiently confident to make the same change
in the Design System.

This is only applied to sort code and account number as the other data collected
is alphanumeric.

Tested on iOS 12 and 13, Android 9 and 8.