Fix sticky navigation logic erroring on some pages
Entries are only appear when an element that is being observed enters or leaves the viewport.

This means that we cannot rely on the subNav entry being avaliable to check it's intersectionRatio.

This change to the logic caches the ratio outside of the observer so it's always up-to-date with the last time an entry for the subNav is avaliable.

This has no impact on users that I can tell fortunately.

Fixes https://github.com/alphagov/govuk-design-system/issues/1088