Add about page with team members
We want to have an official list we can point to for who is on the team.

I considered including our GitHub handles but as long as we have our real names on GitHub you can still cross reference this list.

I've added our job title roles to mirror what we do in internal weeknotes, this is to help people understand what we do on the team.

I've also added a summary of who we are, where we're based, what we're trying to do (based on our mission statement) and how to contact us.