Remove grid overrides
These overrides have been in the Design System since before it was launched, but as far as I can tell the grid issues they are trying to correct were fixed in https://github.com/alphagov/govuk-frontend/pull/332, well before Frontend 1.0 was released.

Importantly, these overrides are also making the grid behaviour inconsistent with other applications using GOV.UK Frontend, as the `-one-third`, `-two-thirds` and `-one-quarter` classes all expand to 100% as soon as the viewport goes below the desktop breakpoint when this should only happen when it goes below the tablet breakpoint.

You can see this best in the example showing all combinations of grid widths together – when you resize to a tablet breakpoint, the three-quarter and the one-half widths both retain their correct proportions, but the others expand to 100%.

This does subtly change the behaviour in a few places in the Design System – notably the masthead on the homepage which is now constrained to two-thirds width on tablet where it wasn't before.