Remove unused classes
`app-content__heading` and `app-content__section-heading` (which used to be `app-c-content__heading` and `app-c-content__section-heading`) have not been used since d09efcf, when they were replaced with the heading and caption classes from GOV.UK Frontend.

`app-content__header` is still being applied, but all it does is set a font-size – as both the caption and the heading have their own explicit font sizes, we can remove this entirely.