Update 'Open in new window' design
Update the 'Open in new window' link so that it sits within its own 'toolbar', rather than overlapping the example.

This allows us to remove the padding from full-page examples without the open in new window link overlapping the header, and simplifies things like the focus state.

It also updates the 'canvas' behind the example (seen only when resizing the example) to a 'checkerboard' design.