Remove unnecessary nav
As this is only just starting with the patterns section and as we want
people to base their work off the central GDS design system then it
makes sense to hide the sections that we do not currently use.

The plan is to add these back in as this matures.