Make the app hostable on Heroku
The `start` npm task tries to run the development version of the site,
which requires that the devDependencies be installed. Heroku prunes
these dependencies upon build, so they aren’t available and trying to
access the site gives an error. So instead, we add a new `serve-static`
task for Heroku to run, which uses the prebuilt site that was built
during the push to Heroku.

The code is pretty much copied from tasks/test-serve.js.